Title: 배열의 특정 위치 찾기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241120/배열의 특정 위치 찾기/find-specific-location-fo-array.c
@@ -3,7 +3,7 @@
 int main() {
     // 여기에 코드를 작성해주세요.
     int n[10], sum=0,sum_1=0, count=0;
-    float avg=0;
+    float avg=0.0;
     for (int i=0; i<10; i++){
         scanf("%d", &n[i]);
 
@@ -13,7 +13,7 @@ int main() {
         if ((i+1)%3==0){
             sum += n[i];
             count++;
-            avg = sum/count;
+            avg = (float)sum/count;
         }
     }
     printf("%d %.1f",sum_1, avg);
